fix : 랭킹 화면 수정
- 스크롤바가 정상위치에서 동작하지 않던 부분 수정
- 월간/주간 탭의 상단바 패딩을 애니메이션 적용

+++ design/src/main/kotlin/com/stepmate/design/component/lazyList/ScrollBarState.kt
@@ -1,5 +1,6 @@
 package com.stepmate.design.component.lazyList
 
+import android.util.Log
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.SideEffect
 import androidx.compose.runtime.derivedStateOf
@@ -35,7 +36,7 @@ class ScrollBarState(
     var threshold by mutableFloatStateOf(maxHeight)
         private set
 
-    val progress by derivedStateOf { offset / threshold }
+    val progress by derivedStateOf { (offset.toDouble() / threshold.toDouble()).toFloat() }
 
     fun onScroll(delta: Float) {
         offset = (offset - delta).coerceIn(0f..threshold)
